refactor: extract ManualTransactionService and move getMonthlyQuotaSum to stats module
- Create ManualTransactionService with create/delete/update/list methods
- Delegate manual TX methods in TransactionService to new service
- Move getMonthlyQuotaSum from TransactionService to StatsService
- Move monthly-sum route from transaction.routes to stats.routes
- Update stats controller with new getMonthlyQuotaSum handler
- TransactionService reduced from ~466 to ~280 lines
